refactor(chat): consolidate ChatInput component into index.tsx
- Merged chat-input.tsx into index.tsx following React best practices
- Removed redundant file structure and self-referential exports
- Updated model-selector-button import to use relative './index'
- Maintains all existing functionality and component exports
- Added 't' dependency to useCallback to fix React Hook warning
